Factor base Illuminator test into its own file for clarity
