+++ Example/IlluminatorUITests/IlluminatorUITests.swift
@@ -10,44 +10,9 @@ import XCTest
 @testable import Illuminator
 import IlluminatorBridge
 
-class IlluminatorUITests: XCTestCase, IlluminatorTestResultHandler {
+class IlluminatorUITests: IlluminatorTestCase {
     
-    // implement IlluminatorTestResultHandler protocol
-    typealias AbstractStateType = AppTestState
-    func handleTestResult(progress: IlluminatorTestProgress<AbstractStateType>) -> (){
 
-        switch progress {
-        case .Failing(let state):
-            print("Failing state was \(state)")
-            // on failure, print out what was on the screen when things failed
-            for line in IlluminatorElement.accessorDump("app", appDebugDescription: app.debugDescription) {
-                print(line)
-            }
-         default:
-            () // do nothing, fall back on default implementation of finish()
-        }
-    }
-    
-    var app: XCUIApplication!
-
-    
-    override func setUp() {
-        super.setUp()
-        
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-        
-        // UI tests must launch the application that they test. Doing this in setup will make sure it happens for each test method.
-        app = XCUIApplication()
-        app.launch()
-        
-        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
     func test_basicWithoutIlluminator() {
         
         let textField = app.otherElements.containingType(.Button, identifier:"Button").childrenMatchingType(.TextField).element
@@ -69,7 +34,7 @@ class IlluminatorUITests: XCTestCase, IlluminatorTestResultHandler {
         initialState
             .apply(interface.home.enterText(myExampleText))
             .apply(interface.home.verifyText(myExampleText))
-            .finish()
+            .finish(self)
     }
 
     func test_basicWithIlluminatorCompositeAction() {
